split out test_chat
