+++ test/src/realtime/test_realtime.cpp
@@ -150,108 +150,6 @@ void test_rt_remote_disconnect()
     test.runTest();
 }
 
-
-void test_rt_joinChat()
-{
-    NTest test(__func__);
-
-    thread::id main_thread_id = this_thread::get_id();
-
-    test.onRtConnect = [&test, &main_thread_id]()
-    {
-        if (main_thread_id != this_thread::get_id())
-        {
-            NLOG_INFO("ERROR: onRtConnect executed not from main thread!");
-            test.stopTest();
-            return;
-        }
-
-        auto successCallback = [&test, &main_thread_id](NChannelPtr channel)
-        {
-            NLOG_INFO("joined chat: " + channel->id);
-
-            if (main_thread_id != this_thread::get_id())
-            {
-                NLOG_INFO("ERROR: successCallback executed not from main thread!");
-                test.stopTest();
-                return;
-            }
-
-            auto ackCallback = [&test](const NChannelMessageAck& ack)
-            {
-                NLOG_INFO("message sent successfuly. msg id: " + ack.messageId);
-                test.stopTest(true);
-            };
-
-            // data must be JSON
-            std::string json_data = "{\"msg\":\"Hello there!\"}";
-
-            test.rtClient->writeChatMessage(
-                channel->id,
-                json_data,
-                ackCallback
-            );
-        };
-
-        test.rtClient->joinChat(
-            "chat",
-            NChannelType::ROOM,
-            {},
-            {},
-            successCallback
-        );
-    };
-
-    test.runTest();
-}
-
-void test_rt_joinGroupChat()
-{
-    NTest test(__func__);
-
-    thread::id main_thread_id = this_thread::get_id();
-
-    test.onRtConnect = [&test, &main_thread_id]()
-    {
-        if (main_thread_id != this_thread::get_id())
-        {
-            NLOG_INFO("ERROR: onRtConnect executed not from main thread!");
-            test.stopTest();
-            return;
-        }
-
-        auto successCallback = [&test](NGroup group)
-        {
-            NLOG_INFO("joined group: " + group.id);
-
-            auto joinedChatSucceeded = [&test](NChannelPtr /*channel*/)
-            {
-                NLOG_INFO("Group chat joined successfully.");
-                test.stopTest(true);
-            };
-
-            auto joinedChatFailed = [&test](const NRtError& err)
-            {
-                NLOG_INFO("Could not join group chat: " + toString(err));
-                test.stopTest(false);
-            };
-
-            test.rtClient->joinChat(
-                group.id,
-                NChannelType::GROUP,
-                {},
-                {},
-                joinedChatSucceeded,
-                joinedChatFailed
-            );
-        };
-
-        test.client->createGroup(test.session, "group chat " + test.session->getAuthToken(), "a group for chatting", "", "", false, {}, successCallback);
-    };
-
-    test.runTest();
-}
-
 void run_realtime_tests()
 {
 //    test_rt_joinChat();
